fix: 학사일정 ID 타입을 Int에서 Long으로 변경
서버에서 Int 범위를 넘는 ID 값이 주어지는 경우가 있어 수정했다. Room은 주어지는 정수 값의 범위에 따라 자동으로 저장 타입을 결정하므로 수정할 필요가 없다.

+++ core/api/src/main/java/com/practice/api/schedule/FakeRemoteScheduleDataSource.kt
@@ -9,7 +9,7 @@ class FakeRemoteScheduleDataSource : RemoteScheduleDataSource {
     private val data = (1..10).map {
         ScheduleModel(
             schoolCode = it,
-            id = it,
+            id = it.toLong(),
             date = Date(2022, it, 10).toEpochSecond(),
             title = "${it}번 학사일정",
             contents = "$it",
